requirements.txt
pinecone
langchain
langchain-groq
langchain-pinecone
langchain-community
sentence-transformers
pypdf2
streamlit
python-dotenv